changed seed data by removing dishes array from country
