+++ server/src/models/Country.ts
@@ -1,15 +1,31 @@
 import { Schema, model, Document } from 'mongoose';
 
+interface IDishes extends Document {
+  name: string;
+}
+
+const dishSchema: Schema = new Schema<IDishes>({
+  name: { 
+    type: String, 
+    required: true 
+  }
+},
+{ toJSON: { getters: true },
+  toObject: { getters: true }, 
+  timestamps: true },
+);
+
+
 interface ICountry extends Document {
   name: string;
-  dishes: Schema.Types.ObjectId[];
+  dishes: [IDishes];
   notes: string;
 }
 
 
 const countrySchema: Schema = new Schema({
     name: { type: String, required: true },
-    dishes: { type: Schema.Types.ObjectId, ref: 'Dishes', required: true },
+    dishes: { type: [dishSchema], required: true },
     notes: { type: String, required: true },
   });
