+++ server/src/models/index.ts
@@ -1,5 +1,5 @@
 import User from './User.js';
 import Country from './Country.js';
-import Dishes from './Dishes.js';
+// import Dishes from './Dishes.js';
 
-export  {User, Country, Dishes};
+export  {User, Country};
