+++ server/src/seeds/countryData.json
@@ -1,41 +1,14 @@
 [
   {
     "name": "Spain",
-    "dishes": [{
-        "name": "Paella"
-      },
-      {
-        "name": "Tortilla Española"
-    }
-    ],
     "notes": "Mediterranean country known for its tapas and Mediterranean cuisine"
   },
   {
     "name": "India",
-    "dishes": [{
-        "name": "Butter Chicken"
-      },
-      {
-        "name": "Biryani"
-      },
-      {
-        "name": "Masala Dosa"
-    }
-    ],
     "notes": "Rich culinary diversity with extensive use of spices and curry"
   },
   {
     "name": "Peru",
-    "dishes": [{
-        "name": "Ceviche"
-      },
-      {
-        "name": "Lomo Saltado"
-      },
-      {
-        "name": "Aji de Gallina"
-    }
-    ],
     "notes": "Cuisine that blends indigenous, European and Asian influences"
   }
 ]
